+++ src/jimdo/gladsoft/anno2016/java/waterquest/quests/EstQuest.java
@@ -1,5 +1,7 @@
 package jimdo.gladsoft.anno2016.java.waterquest.quests;
 
+import android.util.Log;
+
 /**
  * @author fgast34
  * @version 1.0 - 11.07.2016.
@@ -24,19 +26,19 @@ public class EstQuest extends GeneralQuest{
         }
     }
 
-    public EstQuest(float correctAnswer, String desc, String extra) {
+    public EstQuest(double correctAnswer, String desc, String extra) {
         this(correctAnswer,desc, extra, null);
     }
 
-    public EstQuest(float correctAnswer, String desc, String extra, String category) {
+    public EstQuest(double correctAnswer, String desc, String extra, String category) {
         this(correctAnswer,desc,extra,5,category);
     }
 
-    public EstQuest(float correctAnswer, String desc, String extra, float tol) {
+    public EstQuest(double correctAnswer, String desc, String extra, double tol) {
         this(correctAnswer,desc,extra,tol,null);
     }
 
-    public EstQuest(float correctAnswer, String desc, String extra, float tol, String category) {
+    public EstQuest(double correctAnswer, String desc, String extra, double tol, String category) {
         super(desc, extra, category);
         this.mCorrectAnswer = correctAnswer;
         this.mTolerance = tol;
